+++ CHANGELOG.rst
@@ -88,6 +88,7 @@ Changed
     - Refactor ``role_update`` flow usage (#2117)
     - Newline separators in landing zone exception messages (#1308)
     - Do not create timeline events for flows failed by locked project (#1970)
+    - Collect all failed checksums in ``BatchValidateChecksumsTask`` (#1073)
 
 Fixed
 -----